Revert the bit of r326 that I believe was actually causing the Linux issues. (Should negate the need for 354 & 355.)
git-svn-id: http://pcsx2-playground.googlecode.com/svn/trunk@356 a6443dda-0b58-4228-96e9-037be469359c

+++ pcsx2/Misc.h
@@ -123,8 +123,8 @@ typedef struct {
 	int Hacks;
 	int GameFixes;
 	int CustomFrameSkip;
-	u8 CustomConsecutiveFrames;
-	u8 CustomConsecutiveSkip;
+	int CustomConsecutiveFrames;
+	int CustomConsecutiveSkip;
 	u32 sseMXCSR;
 	u32 sseVUMXCSR;
 } PcsxConfig;
